Feat: Improve performance
I had to store and check the source img and the watermark memory references to solve an issue that both images were being re-opened everytime the 'watermark_image' method was called.
Now, it'll store both as instance variables and it'll be cleared only if the file path is changed.
With that, the performance increased a lot when resizing or adjusting the transparency/rotation.